+++ src/core/data-loaders.js
@@ -1,11 +1,15 @@
 class DataLoader {
 
-    async fetchText(url) {
-        return (await fetch(url)).text();
+    async fetchJSON(url) {
+        return (await fetch(url, {
+            headers: { 'x-client-version': "5.1.1" } 
+        })).json();
     }
 
-    async fetchJSON(url) {
-        return (await fetch(url)).json();
+    async fetchText(url) {
+        return (await fetch(url, {
+            headers: { 'x-client-version': "5.1.1" }
+        })).text();
     }
 
     async run() {}
@@ -17,14 +21,13 @@ class ManifestLoader extends DataLoader {
     async run() {
         Manifest = await this.fetchJSON(getResourceURL("manifest.json"));
     }
-    
 
 }
 
 class TipsLoader extends DataLoader {
 
     async run() {
-        ConstantValues._helpfulTips = await this.fetchJSON(ConstantValues.apiURL + "tips" + `?v=${Manifest.version}`);
+        ConstantValues._helpfulTips = await this.fetchJSON(ConstantValues.apiURL + "tips");
     }
 
 }