+++ src/main.js
@@ -11,6 +11,12 @@ let Manifest;
     $("html").css("visibility", "visible");
     Logger.INFO("Extention Starting, Loading Modules")
 
+    if (window.location.href.includes('ban')) {
+        const banLink = ConstantValues.websiteURL + "/banned";
+        alert("Message from chromegle: For more information on why you may have been banned, please visit: " + banLink);
+        return;
+    }
+
     runDataLoaders(
         ManifestLoader,
         TipsLoader